[DRAFT] appc,wgengine: sketch how connectors 2025 hooks into the datapath
This commit outlines basic NAT datapath actions for next-gen
app connectors, and FlowTable structure for caching those actions.
It also demonstrates datapath integration via tstun wrapper hooks,
and presents examples of the methods to be implemented at the
state-management layer (Conn25). It probably should not be merged as
is and this commit message should be re-written. There should be
more detail in the PR description.